+++ crates/ryo3-http/src/headers_like.rs
@@ -15,11 +15,12 @@ pub enum PyHeadersLike {
     Map(PyHttpHeaderMap),
 }
 
-impl From<PyHeadersLike> for HeaderMap {
-    fn from(h: PyHeadersLike) -> Self {
-        match h {
-            PyHeadersLike::Headers(h) => h.read().clone(),
-            PyHeadersLike::Map(d) => d.into(),
+impl PyHeadersLike {
+    #[must_use]
+    pub fn into_header_map(self) -> HeaderMap {
+        match self {
+            Self::Headers(headers) => headers.into_header_map(),
+            Self::Map(map) => map.into(),
         }
     }
 }